Nitpick console logging and rendering
This mostly just improves uniformity and usefulness of logged output
when playing with these objects on the console.

Also, for cases where we have a NullCoordinates object present on a
Cell, we don't want to fail under normal usage. So this adds a few of
the expected methods that Coordinates objects receive. To test:

```ruby
Board.first.cells.first.update(coordinates: nil)
Board.first.render
Board.first.cells.first.render
```

Also, let the board & cells render themselves. This allows for more
fine-grained control of "rendered" output. With this, we add simpler
formatting, which is easier to read when working with (e.g. for a larger
board size).