Automatically update the UI on Cell#reveal
This allows us to "play" the game by randomly revealing Cell objects
on the console. And we don't have to refresh the page anymore either!
Instead, a turbo stream is being used to broadcast a "page refresh"
request via the "boards" channel. We listen for this in the Game - Show
page and Turbo handles automatically refreshing the page for us.
- The "refresh" request is being sent "after_update_commit" on Board,
  which happens due to the new `touch: true` option we added to
  Cell's `belongs_to :board` macro.

Note: redis must be running in order for the broadcast to occur (via
ActionCable).
- Also, we decided to run our "page refresh" broadcasts synchronously.
  We can easily run them asynchronously by changing to
  `after_update_commit  -> { broadcast_refresh_later }` (adding the
  `_later` suffix). But this enqueues a background job and then runs it,
  which just feels like it will be slower. So, until we know otherwise,
  we're going to assume that doing this synchronously won't be a
  problem in production.