Protect against race conditions when creating Games
Add a database uniqueness constraint on Game#status to automatically
handle a race condition where > 1 Game is being created at the same
time. We rescue on this exception type and just `retry`, which will
generally just then go properly grab Game.current this time.

To accommodate this, it's far easier if our initial Game#status is the
same as the in-play ("In-Progress") status. We had previously planned on
not switching the game to In-Progress until the first click + placement
of mines... but this is somewhat arbitrary. We can determine this
pre-mines-placement state by just checking if any mines are present on
the board yet, if needed. But, either way, as soon as the Game/Board
are created, we are actually "in play". i.e. the user can actually
start clicking at any time.

Also, move the logic for finding and finding_or_creating the current
Game into Game. It was an oversight to have this logic in the controller
to begin with. Also, there is no chance of not being able to
find_or_create the current Game so my usual `require_game` before_action
pattern doesn't apply as there's no chance/need for a redirect + flash
notification.