Support flagging Cells / displaying flagged Cells in UI
Flagging is really a "toggle" operation. If the is not currently flagged
then flag it, else un-flag it.

To display the Flag on the UI takes a little more thought...
- A flag is temporary state on the Cell. It should show up as the
  Cell's value if conditions are right. i.e. only if the Cell has not
  yet been revealed.
  - Once a cell is revealed, it's value will be set. And we never want
    to display a flag value once a Cell has been revealed.

To support the above, rename the Cell#deterine_value method to
Cell#determine_revealed_value. This helps make it clear that the
result of this method call only makes sense for revealed Cells. It is
meant to be the "final" value for a Cell. Once set, it should no be
changed...