Redirect based on Game#status on Reveal
If the Game is still In-Progress, we can redirect to the current Game
view. However, if the Game has just ended, we need to redirect to the
Game "show" view, so the User can see the result.

Note: This still causes concern for other players -- those just watching
along or for the ones that didn't make the final reveal that ended a
game. They'll just continue to see the current Game, which will reload
with a new Game once the old Game ends. So they don't get to see the
result of the Game! I think, to satisfy this, attempting to access the
current Game should just redirect to the Game show view from the start.