Add "Reveal Neighbors" function on double click
If a Cell has been revealed and the appropriate number of flags has
been placed in the neighboring cells, then we can save the user time
by just going ahead and revealing all of the non-flagged (and
non-revealed) neighboring cells.

Note: This doesn't guarantee success. If any of the flag are wrongly
placed then a mine could still go off!