Cleanup: Remove "View" from variable names, etc
I've been aching at the feel of the word "View"/"_view" appearing in the
instance variable and local variable names within the view templates.
This is because, after reviewing some of my old practices, I've recalled
a fondness for thinking of my wrapped objects as basically just the
object itself -- just presented in a different way or context. In
general, wrapper objects will forward messages on to the wrapped object.
That may not be explicitly the case with our view objects, because we
don't need that wide of a public API on these, but it is definitely
what happens as needed.

Anyway, I think this reads and feels better in general so trying it out
for now, at least.

Also, fix the scope of the CellViewBehaviors#css_class method. As this
wasn't being referred to by all subclasses, it shouldn't have been
defined within CellViewBehaviors in a way that solely and perfectly fit
the needs of the ActiveCellView class.