Nest CellActions under Board routes/namespaces
The driving factor for this change was noticing how often we had to
build a Result object within the Cell model. This felt right on day 1,
because I was thinking about Cell as a service object for the Board/Cell
actions needed for game play. But thinking from the perspective of a
Cell as just a model, it feels wrong.
- The reason we created the Result object was to feed the Game back to
  the CellActionBehaviors controllers.
  - But we can get the same by just changing the routes a bit. To be
    contextualized by the Board ID, which then has direct access to its
    Game ID.
- Note: Perhaps, if we had already created the (future-planned) Reveal
  service object, the Result object would (will) make more sense.

Note: I chose to move the concern that's now known as
Boards::Cells::ActionBehaviors out of the app/controllers/concerns
folder and into app/controllers/boards/cells because I feel it's better
to have the concern closer to the controllers that are using it. I've
never really been a fan of a separate "concerns" folder for all concerns
vs just calling the concern "[...]Behaviors", as this is signal enough
that the file/class is a concern/mix-in.