Fix double requests on Reveal/FlagToggle/etc
Previously, after adding back the `turbo_stream_from` tag to the top of
the Game Show page, I noticed that each click on a Cell was producing
to separate requests. (Evident in the Network tab in the browser as
well as in the Rails server output.)

This isn't right... and it's not how Radan's game works
(https://minesvshumanity.com). Knowing this, I set out to figure out
why. To make things more similar to Radan's game, I added a link to
the page that would "Reveal a Random Cell". It did not produce double
requests. I'm still not 100% sure why... it could be because the request
produced by the Turbo-intercepted-link click does include a
`turbo-frame` header. Or because of any number of other things that I
just didn't know to reproduce in my JS-only version of sending a POST
request.

So, I decided to just go with a Railsy approach. Which is, I made links
for each of the Cell Action types (Reveal, Toggle Flag, and Reveal
Neighbors), which can be leveraged by our BoardController JS to set the
appropriate `href` and then click the link programmatically.

Further comparing our game with Radan's, I decided to give rendering the
Game via a partial from the Cell Action controllers, as he does. This
proved to be much easier this time around. And, by utilizing the added
`turbo_frame_tag` (wrapping all contents from the "games/_game" partial)
we successfully get:
- a much more minimal response payload for Turbo to then morph the page
  with
- just a single request (with no redirect) when performing Cell Actions

As mentioned, I've also enabled morphing as refresh mechanism for Turbo:
```<%= turbo_refreshes_with(method: :morph, scroll: :preserve) %>```
This should speed things up / make them feel smoother for the end user.
Last time I tried to integrate it, it didn't seem to work as
expected... but I guess I just had more to learn there.

Other issues of note:
- It didn't work to try to put the hidden Cell Action buttons in the
  Board's `<table>` element and then use Stimulus.js's `target` features
  to select them in the UI. This resulted in our Board's `click` events
  firing on these buttons as well since they were then in-scope. And
  caused another kind of "double-submit" issue. So, I reverted that
  and just went with the old `document.getElementById()` instead.

Happy accidents:
- Having to create the Cell Action links gave me the opportunity to
  offer the "Reveal a Random Cell" link (underneath the Game Board).
  And, playing with this, I actually quite like it. Sometimes you just
  *have to* guess in Minesweeper. So this allows the timid to let the
  Database's randomizer pick for them.
  - Note: This same link also serves as the stand-in for performing the
    player-initiated clicks on Cells, as used by the BoardController JS.
    So it serves double duty!