Indicate "loading" by making Cells "ping" on click
Animate the clicked Cell with a "ping" effect. This looks/feels like a
"loading" indicator, in practice. After the board is refreshed, the
animation effect will stop because the cell is replaced with the one in
the normal state.

To facilitate testing this feature (where one needs to find a big, open
space to generate a large cell "pop"), I've also added a "debug" config
option. To use:

1. Start the rails server with `DEBUG=1`:

```bash
DEBUG=1 bin/rails s
```

2. Now, you can just use `Rails.configuration.debug` anywhere in the
   code, where you'd like to see if we're currently in "debug" mode.

References:
- https://guides.rubyonrails.org/configuring.html#custom-configuration