Remove the app.rb initializer b/c it's not necessary anymore
Apparently, Rails 7.1 now autoloads files in the lib directory. See:
https://edgeguides.rubyonrails.org/7_1_release_notes.html#introducing-config-autoload-lib-and-config-autoload-lib-once-for-enhanced-autoloading

As a result, we no longer need to do this ourselves.

I noticed something was amiss when I was working on a new mix-in, added
to the lib directory, but Rails saw it straight away. No need to first
`require` it. Cool!