Add Index View Model for GamesController#index action
This effort started as a desire to move the Game#status_mojis method
out of the Game model and into common View Model code. This has been
accomplished, but it did take some doing.

1. Define a GamesController::Index View Model, and all its sub-View
   Models for handling the needs of the games/index.html page.
   - This lets the games/index.html page talk directly with view models
     all along its journey through listing the past Games. Including,
     calling `status_mojis` method on a view model. Success!
2. In support of the above, add some Behaviors classes to lib.
   - These are commonly used behaviors, based on past experience. Even
     though they're not so commonly used in this app, just yet. But they
     will be if we keep going.

This commit also deprecates the old ViewBehaviorse mix-in in favor of the
new ActiveModelWrapperBehaviors mix-in. ViewBehaviors was overstepping.
It was what we needed for GameView and the like... but those are more
like Presenter objects than Controller-Action-style View Models. And,
e.g., the new GamesControler "Index" View Model is not at all a
Presenter (or even Wrapper) object. So, going forward, we'll split out
these needs into separate mix-ins.