Move Games Index's View Model to app/views/games/index.rb
Trying this out and so far I love it. It puts the view model right next
to the view template that it's serving. Shout out to my friend and
ex-colleague, Kelly, for this idea.

The only trick with this is we have to tell Rails to eager_load from
app/views/**/*. This is done in config/application.rb.
- While I was in there, I noticed the Rails app version was set to 7.1,
  so I changed it to 7.2. I had just missed this during the Rails app
  upgrade.