Migrate remaining View Model code from GamesController
This moves the concern for rendering Games#show out to an independent
Games::Show View Model / file, which is not a small step. Previously,
what we had created what was more like a Presenter object (for Game
models) vs a Controller/Action-oriented View Model. I just messed up a
little bit with the design as it has been a while.

Note: THere are some FIXME's that I'll tackle in a separate step. We
shouldn't be exposing the ActiveRecord models directly to the view, as
we currently must do for generating routes.