Make `router` available to View models
This allows us to render named routes from within View models. Which,
in turn, allows to make access to our ActiveRecord models private in
the context of our View models.

The implementation that allows for the "Router" to work (RailsRouter) is
a Singleton class. Which is: an object that can only ever be
instantiated once. So just a single one of those objects can ever be
built. This single object is built or returned as needed when calling
`instance` on the class. So `Router.instance` is all we ever need to
call to access it, and it is efficient, memory-wise, since only one will
ever be loaded into memory.
- We can then inject this into our methods, which increases flexibility
  to change and decouples the object a bit from this dependency.

Note: While going through these changes, I was getting too confused with
the mix-in names. So I referred to past projects for guidance and
settled on some more illuminating names.

Other changes:
- While coding the new Listing View model, I saw a good opportunity for
  automatic type conversion using a #to_s method, that I wanted to carry
  over into our Cell View model. Basically, when using `<%= %>` in ERB
  templates, whatever you include in them is implicitly type cast into
  a String. We can take advantage of this by defining a #to_s method and
  then we can therefore just output our object "directly". e.g.
  `<%= cell %>` (instead of `<%= cell.render %>` or even
  `<%= cell.to_s %>`.
- Updated the inner-workings of some of the new mix-ins, now that we're
  getting some play time in with them. e.g. ActiveModelWrapperBehaviors.
  - On this one, I decided it's best not to assume the instantiation
    method's behavior, and just require including objects to define it
    for themselves.
- Fix the scopes for some of our methods to be private, etc.