Show indicator onClick of Reveal Random Cell button
When clicking cells on the board we now show a loading indicator, in
case of a big pop of cells that takes a while to show up. But, the same
wasn't yet true for use of the "Reveal a Random Cell" button.

This commit adds that same benefit to use of this button.

Also:
- Fix the magic value `0` for the other 2, hidden buttons: Toggle Flag
  and "Reveal Neighbors". Both of these need to use some value that
  isn't a valid Cell ID.