Abstract out the concept of "Difficulty Level"
Previously, Board held on to the concept of "Difficulty Level" via a
constant called DIFFICULTY_LEVELS_MAP. This was OK at first, but
arguably belongs more to the Game than to the Board. Plus, it's a bit
of a popular subject. It helps to identify the type of Game being
played, of having been played (in the past) for example.
- This analysis started by me wanting to indicate the difficulty level
  of past games in the Sweep Ops Archive view.
- I also noted that we were using basically this same hash to determine
  what buttons to render in the "New Game" view, directly. Of course,
  I wanted to change this page to be controlled by a View Model as
  well...

In service of the Single Responsibility Principle, then, and because I
didn't want to add more logic to Board for displaying Difficulty Levels,
I set out to create my own PORO. From which the new DifficultyLevel
object comes!