Take / Convert to a DifficultyLevel object in Game.build_for
Now that we have the DifficultyLevel object, this is the concept that
names what we were passing in as keyword arguments before: columns,
rows, mines. So, instead, we can now pass in an actual DifficultyLevel
object for consumption here.

Of course, it's not super duper easy to just "have" a DifficultyLevel
object, e.g. on the Console. So we can have our cake and eat it to by
defining a Conversion Function to handle converting whatever we do get
to a DifficultyLevel object, as best we can. This adds flexibility to
other object types in the future (Open/Closed principle), and even
reduces coupling on the Game.build_for_custom method.
- Note: It also gets rid of the LongParameterList code smell that reek
  was warning us about before. (And this is how we do it... when we
  have a long parameter list, we reduce it by naming the concept and
  passing it in as an object instead.)

Also:
- Make usage consistent between Game and Board as far as what argument
  type (positional or keyword) `difficulty_level` is expected to be.
  It's more to type on the console, but consistency is worth it.