Add ability to start a Game at a randomized Difficulty
We can now show a "Random" button in the New Game view.

The only thing difficult about this is that we need actual
DifficultyLevel-like objects to stand in for themselves when building
the view (via the Games::New View Model). So we need a stub object to
play this role (and nothing more). From there, we can rely our new
Conversions.DifficultyLevel to interpret the selection (by name) and
then actually do the picking of a random DifficultyLevel object to
build for consumption by the back-end.