Add a Timer / Stopwatch to the Game#show view
This solution implements an ElapsedTime object/behavior twice. Once in
Ruby and once in JavaScript.

The Ruby implementation is necessary because we don't want a flash of
time when the page first loads / reloads where there is no timer
displayed. Especially since we're doing constant morphing of the page --
each time a player interacts with a Cell.

The JavaScript implementation is necessary because as a user just sits
on the page, we want the timer to count upwards on its own. And we don't
want to have to poll the server over and over again.

Any potential disparity between the time calculated on the server vs the
timer value as calculated on the front-end will be brief/small. Because
each Cell interaction will refresh the page with the latest from the
server (which is the authoritative source).

Note: We're at the mercy of the Game#created_at and Game#updated_at
attributes with this implementation. But... that's OK. We can make this
more robust in the future if/when we add Transaction records for events
such as Game start/stop.

Noteworthy features:
- We are careful to stop the timer on Game end, as well as on disconnect
  of the Stimulus controller. This helps to prevent memory leaks.
- Once the game is over, we'll no longer register the
  Stimulus controller (StopwatchController). This makes display of the
  Timer possible even on past Games.
- We max out our Timer at "23:59:59". Any value above this will simply
  show as "23:59:59+". This saves us the hassle of tracking Games that
  run on for days, weeks, months, or even years.
  - I started out trying to track time beyond just a single day, but it
    really was a lot of extra work... and it just won't generally be
    of any value. So, it wasn't worth keeping that around.