Redirect to the current game if any from Game#new action
Sometimes, due to errors I encounter or whatever in development, I end
up refreshing the Games#new page, while there is already an current
game in progress. Whatever the reason, though, if there's a current
game in progress, it does no good to be trying to start a new one.