Adjust Game statuses; Add "Standing By" as preliminary
We had originally tried to initialize games in a special status called
"Initializing", and then transition them to "In-Progress". Ensuring
uniqueness between both of these statuses proved difficult so we just
abandoned the approach at the time. But, we need it now.

Why? Because we don't want to:
- place mines
- start the Game timer
... until the Game has "Started" started (no for realsies!).

So, alas, we need more states (statuses) to key off of!

This commit, therefore:
- adds a preliminary status: "Standing By"
- renames the existing "In-Progress" status to "Sweep in Progress"
... and then cleans up all the ramifications of this.

Note: While both of these preliminary statuses are arguably
"in progress"-like, one is more in progress-er than the other. So the
terms I've settled on to make this more clear here are:
- "Game On" = either status: Standing By or Sweep in Progress
- "Game Over" = neither of these statuses; i.e. either: "Alliance Wins"
  or "Mines Win"
- And "Game In Progress" = status: Sweep in Progress

To accomplish the database constraint for preventing more than one Game
record from existing if another is already in either of the to "Game On"
statuses required writing a trigger. As I'm not so great with this
stuff, our trigger comes courtesy of Grok. It took a few iterations, but
this result does seem to do the trick.
- To accommodate creation of the database trigger, I chose to just
  rewrite our original CreateGames migration since we're still not to
  production yet and since this is a major time saver. However, it does
  require dropping the database and recreating it in development envs.

Other:
We also take a small step here towards the goal of not placing mines
until the games transitions to stay Sweep in Progress by adding a
Game#start method. But, at this point, functionality remains the same as
it was. i.e. Game creation immediately moves the Game into
"Sweep in Progress" status and places mines, without delay.