Make the first click (reveal) of the Game always safe!
(As it should be, per MS Minesweeper since forever.)

This makes use of our new Standing By vs Sweep in Progress statuses. If
the Game is Standing By then we want to "start" the game by:
1. Placing mines at random, except explicitly not in the Cell that was
   just revealed
2. Changing the Game status to Sweep in Progress

Game Status-based checking prevents us from having to scan all of cells
for any mines to know if mines need to be placed or not. And since we
already have the associated Game object available to us in the
Games::Boards::Cells::RevealsController, this change was almost no
sweat.

Gotchas:
- When recursivly revealing neibhgoring cells, we don't really want to
  have to pass the current Game object along as well. It's not needed
  since by the time we get to the recursive reveal we will have already
  started the Game. So, we choose to just not pass along our Game object
  in this case and instead deal with a nil type check for the call to
  Game#start.

Note: A great way to determine this feature is working to use DEBUG
mode, which allows us to see which cells contain mines. Start DEBUG
mode with: `DEBUG=1 bin/rails s`. Further, checking the Rails console
after clicking the first cell shows that the DB query does indeed
exclude Cells with the ID of the clicked cell.