Show Game Engagement Time Range on Sweep Ops Archive pg
Previously, we were only showing the Game#updated_at (which was supposed
to represent the end time) for reference. But now, it's pretty easy to
show the start-end time range, so let's do it.

Caveat: We now require more horizontal space for display details about
each of the past games. So, we have to reduce the number of columns we
can show at once to accommodate this change.