Redo: Show a message on the Games#index pg when no listings
This is a re-do of the previous commit with the same subject:
eb0666a335910215dc3b070db6acd245e03eee7b

Instead of always rendering the "empty" message, we'll just use a
conditional. The always-render approach makes more sense when it's
possible to gain as well as lose list items, as a matter of course (like
with a TODO list). In our case, it's either we have Games to list or
we don't.