Don't send reveal_neibhgors POST for blank Cell clicks
If a cell is blank, then it all of its neighbors will already be
revealed. So, going through with sending the POST request to the server
is pointless. This was just an oversight before. I noticed it while
debugging the JS code for the next big feature.

While I was in there, I decomposed (abstracted) some of the repeated
code/concepts as well.