Highlight neighboring cells of revealed cells onMouseDown
<SPOLIER ALERT>All of this code will end up getting removed in the next
commit. But I wanted to keep it around for future reference / just in
case.</SPOILER_ALERT>

This adds another effect from the MS game that really helps when you
start going cross-eyed from looking at all the tiny little squares. i.e.
While pressing down on the mouse, highlight all unrevealed/unflagged,
neighboring cells. On mouse up, it may be the case that this will also
auto-reveal those cells if the proper conditions have already been met.
But, if not, then this is quite helpful for visualizing the current
cell's radius better.

This required re-implementing the Ruby logic for finding neighboring
cells into JS, which is a bummer. The effect is worth it!... but now I
realize I probably could have just done this with Turbo. Doh!

Note: As is, we never undo the Cell highlight effect... on purpose.
Why? Because, onMouseUp, the "Reveal Neighbors" POST request fires and:
- either it is valid, which reveals neighboring cells and refreshes the
  board
- or, it is invalid, which still refreshes the board. (Which clears our
  CSS style changes for us.)