Redo: Highlight neighboring cells of revealed cells onMouseDown
This adds another effect from the MS game that really helps when you
start going cross-eyed from looking at all the tiny little squares. i.e.
While pressing down on the mouse, highlight all unrevealed/unflagged,
neighboring cells. On mouse up, it may be the case that this will also
auto-reveal those cells if the proper conditions have already been met.
But, if not, then this is quite helpful for visualizing the current
cell's radius of influence.

My first attempt at this (see previous commit) was a bust. I did the
whole thing in JavaScript. But then, while typing out the commit
message, I realized that I shouldn't have to implement this logic in JS
any more than for flagging, revealing, etc. The whole point of
Turbo/Stimulus is to keep our view-layer interaction as HTML Over The
Wire! It just... seemed absurd to think about at first. Could it really
be fast enough ... to be able to respond to an onMouseDown/onMouseUp
event via 2 separate server requests?! Spoiler: Yes. Wow.
- I think we're benefiting greatly from Turbo 8's DOM morphing logic
  here. This allows us to preserve the board + un-altered Cells and,
  thus, properly capture the mouseup event.

Note: We lean on our existing Reveal Neighbors hook to undo the Cell
highlight effect. This is not only convenient, but it makes good sense,
too. Why? Because a reveal attempt is basically the same thing as a
highlight, it just actually does the reveal work if the conditions are
appropriate. So, we can just make highlight/reveal 2 sides of the same
coin and everyone wins.