Add "Reveal[Neighbors]" Service Objects
Previously, we put a lot of logic into Cell#reveal, outside of just
doing the actual Cell reveal. This was a shortcut to what is included
in this commit, now.

The new Reveal Service Objects can easily handle the complexities of
collaborating between Game, Board, and Cell as needed, because they are
neutral 3rd-parties to all of these objects.

Note: The heart of the Reveal Service Objects are the #on_call methods,
which are simple "composed methods". i.e. they compose the flow of the
Service Object (in basically readable English), without concerning
themselves with any conditionals or other logic branches.

See the comments on the Service Objects themselves for in-depth info on
what is happening with the changes in this commit.
- Note: I'm trying out just defining these service objects on the
  Controllers that use them. They are only usable/used in these spots
  anyway, as it is right now. If that changes, we can move them out
  to a `services` folder or something.

Also:
- Fix a bug in the CallMethodBehaviors code. I guess we hadn't actually
  gotten to use this yet, at least with any objects that expect keyword
  arguments. Passing on method arguments with `(...)` is the Ruby 3.x
  way.

References:
- On `throw` / `catch`:
  https://avdi.codes/throw-catch-raise-rescue-im-so-confused/