Show Cell#[de]highlight to all players via TurboStream
I had assumed that `update_all` would fire the callback to touch the
Board for highlight/dehighlight, but that was wrong. `update_all`
doesn't fire any callbacks or instantiate the objects. It does return
the numbers of rows affected, however. So we can just manually touch the
associated Board conditionally--when there were any affected Cells.

This allows other players to see when someone is highlighting other
cells by clicking & holding on a revealed Cell. Makes the game feel more
interactive.