Tidy up naming of `is_mine` and `is_not_mine` scopes
It sounds better to say `is_a_mine` and `is_not_a_mine`, but this goes
against my convention, so it's harder to remember what is real. e.g.
all the other flags are just `is_<flag_name>`. So it's unclear if the
flag (boolean attribute) is actually `a_mine` or not.