Introduce the concept of *::Console objects
This commit involves rearranging each of our model objects to:
- Reduce the scope of its public API
- Move IRB Console-only code out of the ActiveRecord Model, directly

The idea is that while it's convenient to have codified access to
IRB Console-friendly helpers and utility methods on our models, it's
also "bad" to define many of these as being part of the model's public
API.
- Methods in the public API must be maintained at a higher standard of
  integrity (including Testing) than even private methods, but
  especially more so than those that are just meant to be
  IRB Console-specific helpers.
- Some IRB Console-specific methods could be considered dangerous to
  (accidentally) run e.g. in production (but perhaps even in
  development, to an extent).

So, we can add a `console` method to participating models, which will
return an IRB Console-specific wrapper of that model. The *::Console
objects will present themselves as if they are the original model, and
they will respond to all the same messages... but they can be more, too.

One obvious example is Object Inspection. I brought in my
ObjectInspector gem to help with formatting objects for inspection. And,
this is great... but maybe we don't always want to format objects this
way. Sometimes, it can even cause extra loading requirements to do so.
(This is why the gem introduces `scopes`, but that's beside the point
here.) Either way, for the *::Console objects, we'll use the pretty
inspect Strings, as put together by our use of ObjectInspector, and the
non-*::Console objects will not.

Another good example is how I've now moved Board#reset to
Board::Console#reset. "Reset" is a "dangerous" operation that we
shouldn't be calling unless we definitely intend it. To signal
intention, we now must do:

```ruby
Board.last.reset  # => (NoMethodError)

Board.last.console.reset
Board.last.cons.reset
```

I've also moved all of our `render` methods/logic into the *::Console
objects, as these were made purely for inspection in the IRB Console in
the first place.