Display Flags/Mines counts instead of just Flags Remaining
I think this makes more sense... instead of implying how many flags
remain to be set--as if it's known to be correct, we can just display
how many Flags have been set vs how many Mines there actually are on
the board.

Also, doing this allows us to remove the "total mines count" indicator
next to the difficulty level info at the bottom-right of the Board.

Further, prevent the displayed items in the <thead> from running
together and/or word-wrapping on smaller Board sizes, such as for the
"Test" Difficulty Level. (Not that this is a realistic Difficulty Level
for production use.)