Add ability to reset a Game on the console
This is a debug measure... but it allows us to reset an entire Game
(keeping the mines where they were) in case we see something weird while
testing.

Note that this is only "possible" on the IRB Console, via our new
Game::Console and Board::Console objects.

To use:

```ruby
Game.last.console.reset
```

Note that this will automatically refresh the Home page with the
freshly-reset Game as well.

Other changes:
- Add *::Console::Error class for use as needed, via the
  ConsoleObjectBehaviors mix-in.
  - Which we use here, in Game::Console#reset: `raise(Error, "...")`