Prefer broadcasting refreshes manually vs: via callbacks
Callbacks are "evil". They get you into trouble down the road. They
make testing harder, for one thing. And anything that makes testing
harder reveals a problem with the design. i.e. with using callbacks.

The alternative is to just manually call the thing as needed. During
processing of service objects, etc.

Doing this also allows us to get rid of the Board#touch every time a
Cell updates. That's a fair amount of overhead reduction. Even if it may
not add up to much in actual processing time, it at least reduces
clutter in the Rails server logs and *feels* better 😜.

Note: I don't necessarily love making `broadcast_changes` part of the
controller actions... but it's at least an easy, uniform spot to do
this, for now. We don't yet have Service Objects for all 4 of the Cell
Actions controllers, so it wasn't feeling uniform to put them into
the Service Objects of only those that do. Can reconsider later.

References:
- https://blog.corsego.com/turbo-hotwire-broadcasts#8-best-practices-when-broadcasting