Indicate which is the the Currently Active Nav Link
This uses one of my long-time helper methods for adding an `active` CSS
class to the currently active main Nav Links.

Then we just style what it means to be an `a.active` with CSS.

Also: Add a bottom border to the header section for visual break-up
from the rest of the page.