Sort out Reek offenses and warnings
This trims our number of warnings coming from Reek down from about 10
to just 1. The 1 remaining warning is legitimate, and I plan to do
something about it:
- TooManyMethods: Cell has at least 17 methods
  https://github.com/troessner/reek/blob/v6.3.0/docs/Too-Many-Methods.md

I'm still having trouble with SublimeTest not respecting the project's
.reek.yml config file, so sometimes I disable a warning that isn't
needed. Unfortunately, Reek doesn't call this out, so I don't always
know. So one or 2 of the changes in this commit is just me removing an
unneeded disabling of a warning.

Issues I'm taking a stand on:
- I like calling `Icon.<icon_name>`, even if multiple times in one
  method. Plus, I like calling it in the view templates. I thin it's
  quite clear and simple, and I don't need the View Model stepping in
  to provide a convenience method to do the same. (So. Seeing how this
  goes for now haha.)
- I will never allow a method or parameter or variable name of `x` or
  `y` except in the cases used in this project: for X/Y Coordinates.
  So, I'm just disabling these detectors.
- Allow duplicate method calls to `self.class`, because that's just how
  it be, sometimes.
- I dislike reek's UtilityFunction detector in general... but I'm not
  going to disable it because sometimes it's right and I should pay
  attention to it. So I'll continue to just disable it in many of my
  private methods (sigh).