Show current number of connected players in Game status
NOTE: This is a bit crude, and also likely buggy... but still better
than nothing?

This is an attempt to show the current number of user sessions that
are connected to the Turbo Stream "current_game". It relies on a monkey
patch that I've added to the core Turbo::StreamsChannel code to
increment/decrement the current number of players (a.k.a. the Duty
Roster) that we currently know about.
- This is the buggy part! Though it may just be due to my Rails server
  being in development mode and doing code reloading, etc.

We store the count of the number of current players in the Rails.cache
memory store. This works well as we would never want to persist this
count beyond a server shut down / restart.

Notes:
- The number of currently connected players is displayed somewhat
  subtly... as the number of ships that are rendered next to the current
  game status. This is meant to represent the number of Minesweepers
  reporting for duty.
- Since there is no Turbo Stream channel setup on past games, the view
  there should not be affected. Also, it should not affect the current,
  active Game. i.e. casually browsing the site won't affect the
  DutyRoster.
- Navigating away from the current game does properly decrement the
  Duty Roster, but then coming right back doesn't seem to increment it.
  This is due to the TurboStream subscribe/unsubscribe events not
  properly firing, I guess. And, again, may just be development
  environment-only issue 🤷
- To enable the Rails Cache Store in development environment, you must
  run: `rails dev:cache`. This command toggles caching on/off (in the
  development environment only).