Remove `lint` task from GitHub Action: CI
It's... buggy? It runs for hours, scanning every single line of ruby
source code and all gems in the project... before finally failing.
I'm unable to find anyone else having the problem, but this is the code
straight from Rails 7.2 release. So... just disabling it for now and
will try again some other day.

Also, add a CI status badge to the README. Furnished by GitHub Actions.
Because why not.