Move Game creation broadcast to controller
Why?
- I felt the pain when trying to write a test for Game.create_for, which
  previously included the Turbo Streams broadcast. I listened to the
  pain :)
- It occurred to met hat Controller actions like `create` are almost
  like little, unwrapped Service Objects. So it makes sense to move the
  orchestration behavior (sending a broadcast on creation) to the thing
  that is responsible for the "event" of creation. i.e. the controller
  action or service object, whichever is available.

Either way, doing this gave me the confidence to finish the Game
unit tests. Specifically the Game.create_for test example.

Also, it made me realize that we had a bug in our GamesController#create
action with clearing the DutyRoster. This should only happen when a new
Game was just created within this action. Because we could otherwise
have just gotten the currently active Game, in which case we don't want
to clear out the known-players list.

Other changes:
- Added ApplicationRecord#just_created? as a friendlier and more
  intention-revealing method than `id_previously_changed?`. In practice,
  this tells us if the game was already active (Game.current) or if we
  just created a new one, since Game.find_or_create_current doesn't
  tell us which it did.