Add double-submit protection to "Reveal Random" button
The Reveal Random button is a bit special. There is no form-submission
happening when you click it. At least not directly. We could have made
it this way, but then we would have had to re-implement all the logic
that happens when clicking a Cell (to make it show the pulse animation,
do the form submission, etc.). So, as a result, clicking Reveal Random
doesn't disable the button as is normal for Turbo form submissions.

So, we just need to disable it ourselves. Which is quite easy.

Also, style disabled primary-buttons so that they keep their red
border/ring. This way they don't flash "off", as if they're no longer
pressed while waiting on the form to submit. (It helps to add a `sleep`
statement in the Rails controller to see this effect in action.) It
applies to all `button_to` elements (which normally do disable
themselves on submit), including the Difficulty Level /
New Mission buttons when creating a game.